+++ chain/index.js
@@ -1,11 +1,11 @@
+const broadcastTransaction = require('./broadcast_transaction');
 const getMaxFundAmount = require('./get_max_fund_amount');
 const getNetwork = require('./get_network');
 const getTransactionRecord = require('./get_transaction_record');
-const getTransitRefund = require('./get_transit_refund');
 
 module.exports = {
+  broadcastTransaction,
   getMaxFundAmount,
   getNetwork,
   getTransactionRecord,
-  getTransitRefund,
 };