+++ funding/index.js
@@ -1,0 +1,11 @@
+const getFundedTransaction = require('./get_funded_transaction');
+const getTransitRefund = require('./get_transit_refund');
+const maintainUtxoLocks = require('./maintain_utxo_locks');
+const reserveTransitFunds = require('./reserve_transit_funds');
+
+module.exports = {
+  getFundedTransaction,
+  getTransitRefund,
+  maintainUtxoLocks,
+  reserveTransitFunds,
+};