+++ index.js
@@ -1,8 +1,10 @@
 const {acceptsChannelOpen} = require('./peers');
+const {broadcastTransaction} = require('./chain');
 const {connectPeer} = require('./peers');
 const {findKey} = require('./peers');
 const {formatTokens} = require('./display');
 const {getAllInvoices} = require('./transactions');
+const {getFundedTransaction} = require('./funding');
 const {getLiquidity} = require('./peers');
 const {getMaxFundAmount} = require('./chain');
 const {getNetwork} = require('./chain');
@@ -12,16 +14,21 @@ const {getPeerLiquidity} = require('./peers');
 const {getRebalancePayments} = require('./transactions');
 const {getScoredNodes} = require('./graph');
 const {getTransactionRecord} = require('./chain');
-const {getTransitRefund} = require('./chain');
+const {getTransitRefund} = require('./funding');
+const {maintainUtxoLocks} = require('./funding');
+const {reserveTransitFunds} = require('./funding');
+const {stopAllHtlcs} = require('./peers');
 const {updateChannelFee} = require('./peers');
 const {waitForPendingOpen} = require('./peers');
 
 module.exports = {
   acceptsChannelOpen,
+  broadcastTransaction,
   connectPeer,
   findKey,
   formatTokens,
   getAllInvoices,
+  getFundedTransaction,
   getLiquidity,
   getMaxFundAmount,
   getNetwork,
@@ -32,6 +39,9 @@ module.exports = {
   getScoredNodes,
   getTransactionRecord,
   getTransitRefund,
+  maintainUtxoLocks,
+  reserveTransitFunds,
+  stopAllHtlcs,
   updateChannelFee,
   waitForPendingOpen,
 };