+++ peers/index.js
@@ -2,7 +2,8 @@ const acceptsChannelOpen = require('./accepts_channel_open');
 const connectPeer = require('./connect_peer');
 const findKey = require('./find_key');
 const getLiquidity = require('./get_liquidity');
-const getPeerLiquidity =  require('./get_peer_liquidity');
+const getPeerLiquidity = require('./get_peer_liquidity');
+const stopAllHtlcs = require('./stop_all_htlcs');
 const updateChannelFee = require('./update_channel_fee');
 const waitForPendingOpen = require('./wait_for_pending_open');
 
@@ -12,6 +13,7 @@ module.exports = {
   findKey,
   getLiquidity,
   getPeerLiquidity,
+  stopAllHtlcs,
   updateChannelFee,
   waitForPendingOpen,
 };