+++ test/integration/test_get_node_alias.js
@@ -1,5 +1,5 @@
-const {test} = require('@alexbosworth/tap');
 const {spawnLightningCluster} = require('ln-docker-daemons');
+const {test} = require('@alexbosworth/tap');
 
 const {getNodeAlias} = require('./../../');